% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvt-package.R
\docType{package}
\name{acvt-package}
\alias{acvt}
\alias{acvt-package}
\title{acvt: Academic-CV Typst Template for Quarto}
\description{
Create academic CVs and cover letters using the Academic-CV Typst template with this Quarto Extension. Includes an RStudio Project Template.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/orehren/acvt}
  \item Report bugs at \url{https://github.com/orehren/acvt/issues}
}

}
\author{
\strong{Maintainer}: Oliver Rehren \email{o.rehren@gmail.com} (\href{https://orcid.org/0000-0001-7602-8219}{ORCID})

}
\keyword{internal}
