% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvt_template.R
\name{acvt_template}
\alias{acvt_template}
\title{Create a new Academic-CV Project}
\usage{
acvt_template(path, ...)
}
\arguments{
\item{path}{The path to the new project.}

\item{...}{Dynamic arguments passed by RStudio based on the .dcf file.}
}
\description{
Implements the logic for the RStudio Project Wizard.
Accepts dynamic arguments via ellipsis.
Adds a temporary self-destructing hook to open the main file on first launch.
}
