% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.classify_source}
\alias{.classify_source}
\title{Classify Source Type based on File Extension}
\usage{
.classify_source(path)
}
\arguments{
\item{path}{File path or Google ID.}
}
\value{
List with \code{path} and \code{type}.
}
\description{
Classify Source Type based on File Extension
}
