% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.expand_path_to_sources}
\alias{.expand_path_to_sources}
\title{Expand Single Path to Source Objects}
\usage{
.expand_path_to_sources(path)
}
\arguments{
\item{path}{String path or ID.}
}
\value{
List of source objects.
}
\description{
Expand Single Path to Source Objects
}
