% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.find_fuzzy_match}
\alias{.find_fuzzy_match}
\title{Fuzzy Match String in Candidates}
\usage{
.find_fuzzy_match(target, candidates)
}
\arguments{
\item{target}{String to find.}

\item{candidates}{Vector of strings to search in.}
}
\value{
Matching string from candidates or NULL.
}
\description{
Fuzzy Match String in Candidates
}
