% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.get_conf_val}
\alias{.get_conf_val}
\title{Helper to safely retrieve config values with defaults}
\usage{
.get_conf_val(config, key, default)
}
\arguments{
\item{config}{List object.}

\item{key}{String key.}

\item{default}{Fallback value.}
}
\value{
The value or the default.
}
\description{
Helper to safely retrieve config values with defaults
}
