% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.get_quarto_config}
\alias{.get_quarto_config}
\title{Retrieve Configuration from QMD Frontmatter}
\usage{
.get_quarto_config()
}
\value{
A list containing the configuration or NULL if not found.
}
\description{
Retrieve Configuration from QMD Frontmatter
}
