% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.merge_source_lists}
\alias{.merge_source_lists}
\title{Merge Two Data Lists}
\usage{
.merge_source_lists(existing_list, new_list)
}
\arguments{
\item{existing_list}{Accumulator list.}

\item{new_list}{New list to merge in.}
}
\value{
Merged list.
}
\description{
Merge Two Data Lists
}
