% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.normalize_sheet_config}
\alias{.normalize_sheet_config}
\title{Normalize Sheet Configuration}
\usage{
.normalize_sheet_config(raw_config)
}
\arguments{
\item{raw_config}{List from YAML.}
}
\value{
A named list: names = Real Sheet Names, values = Short Names.
}
\description{
Normalize Sheet Configuration
}
