% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.read_flat}
\alias{.read_flat}
\title{Read Flat Files (CSV/JSON)}
\usage{
.read_flat(path, type, sheets_map)
}
\arguments{
\item{path}{File path.}

\item{type}{"csv" or "json".}

\item{sheets_map}{Config map.}
}
\value{
Named list of DataFrames (Keys = Short Names).
}
\description{
Read Flat Files (CSV/JSON)
}
