% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.read_google}
\alias{.read_google}
\title{Read Google Sheet}
\usage{
.read_google(id, sheets_map, email)
}
\arguments{
\item{id}{Google Sheet ID or Name.}

\item{sheets_map}{Config map.}

\item{email}{Auth email.}
}
\value{
Named list of DataFrames.
}
\description{
Delegates loading to the helper script.
}
