% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.read_source}
\alias{.read_source}
\title{Dispatcher for Source Reading}
\usage{
.read_source(source, sheets_map, auth_email)
}
\arguments{
\item{source}{Source object.}

\item{sheets_map}{Config map.}

\item{auth_email}{Email string.}
}
\value{
Named list of DataFrames.
}
\description{
Dispatcher for Source Reading
}
