% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.read_xls}
\alias{.read_xls}
\title{Read Local Excel File}
\usage{
.read_xls(path, sheets_map)
}
\arguments{
\item{path}{File path.}

\item{sheets_map}{Config map.}
}
\value{
Named list of DataFrames (Keys = Short Names).
}
\description{
Read Local Excel File
}
