% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.resolve_sources}
\alias{.resolve_sources}
\title{Resolve Source Identifiers to Source Objects}
\usage{
.resolve_sources(raw_ids)
}
\arguments{
\item{raw_ids}{List or vector of identifiers.}
}
\value{
A list of lists, each containing \code{path} and \code{type}.
}
\description{
Resolve Source Identifiers to Source Objects
}
