% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.robust_bind_rows}
\alias{.robust_bind_rows}
\title{Robust Row Binding}
\usage{
.robust_bind_rows(df1, df2)
}
\arguments{
\item{df1}{DataFrame 1.}

\item{df2}{DataFrame 2.}
}
\value{
Combined DataFrame.
}
\description{
Robust Row Binding
}
