% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cv_data.R
\name{.write_json_cache}
\alias{.write_json_cache}
\title{Write JSON Cache}
\usage{
.write_json_cache(data, path)
}
\arguments{
\item{data}{Data to save.}

\item{path}{File path.}
}
\description{
Write JSON Cache
}
