% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_current_cv_project}
\alias{update_current_cv_project}
\title{Update Academic-CV Extension Files in Current Project}
\usage{
update_current_cv_project()
}
\value{
Invisible NULL.
}
\description{
Updates the \verb{_extensions} folder in the current project to match the version
currently installed in the \code{acvt} R package.
}
\details{
Use this function after updating the \code{acvt} package to apply the latest
styles and logic to your existing CV projects.
}
