% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_template_from_git}
\alias{update_template_from_git}
\title{Update Installed Template from GitHub}
\usage{
update_template_from_git(repo = "orehren/acvt", branch = "main")
}
\arguments{
\item{repo}{The GitHub repository in "user/repo" format.}

\item{branch}{The branch to download from. Defaults to "main".}
}
\value{
Invisible NULL.
}
\description{
Downloads the latest template files (skeleton) from the GitHub repository
and overwrites the files in your installed 'acvt' package library.
}
\details{
This allows you to get the latest fixes for the CV template and extension
files without needing to reinstall the entire R package.
}
