% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvt_template.R
\name{acvt_template}
\alias{acvt_template}
\title{Create a new Academic CV Project}
\usage{
acvt_template(path, firstname, lastname, renv, git, ...)
}
\arguments{
\item{path}{The path to the new project.}

\item{firstname}{The user's first name.}

\item{lastname}{The user's last name.}

\item{renv}{A boolean indicating whether to initialize renv.}

\item{git}{A boolean indicating whether to initialize a git repository.}

\item{...}{Additional arguments.}
}
\description{
Implements the logic for the RStudio Project Wizard.
}
