% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_project_template.R
\name{update_current_cv_project}
\alias{update_current_cv_project}
\title{Update CV Extension Files in Current Project}
\usage{
update_current_cv_project()
}
\value{
Invisible NULL.
}
\description{
Updates the `_extensions` folder in the current project to match the version
currently installed in the `acvt` R package.
}
\details{
Use this function after updating the `acvt` package to apply the latest
styles and logic to your existing CV projects.
}
